+++ changelog.md
@@ -7,3 +7,4 @@
   * Renamed `small_hedges` tags into `hedges`
 * Improved hedges block and model to make them more visually consistent (Hedges already placed will need to be updated)
 * Hedges can now connect to leaves
+* Added Light Gray Stone Bricks for consistency (slightly lighter than stone bricks)